Add function to wrap last link word in span
Currently we are manually adding `span` tags around the text content of
links and buttons to prevent the chevron that follows them from wrapping
on its own. This is a chore, error prone and leads to untidy markup
that's difficult to search.

This is the first step in remedying it; elements that have the `.button`
class will, when the page is loaded, have the final word of their
content automatically wrapped in a span.