Support skipping optional steps that get pre-filled
If a candidate is matched back to an existing contact in the CRM we pre-fill
the form fields where possible.

Currently, if the user lands on an optional step that is pre-filled with a
value and they then clear the value the change won't be reflected in the CRM,
as the API only supports sending non-nil values to the CRM.

To make the user-experience consistent for now we are going to skip optional
steps that get pre-filled. A more robust solution (which is also a lot more
work) would be to update the API to support PATCH requests so that partial
updates could be performed that contain `nil` values. The difficulty with this
is that the `swagger-codegen` doesn't look like it will generate a client that
supports it, but I'm going to investigate it further.