Add intro to mailing list sign up via magic link
Records the `auth_method` in the wizard store, so we can later determine if
they used an access token or magic link token to authenticate.

Adds copy to the `DegreeStatus` step to give an intro/context to the sign up
when the user landed straight on this step via a magic link.