Support hidden and optional wizard steps
Currently if a step is `skipped?` it will not show in the
wizard and the data will not be exported to the CRM.

We are introducing `optional?` steps, whereby if they are
pre-filled from the CRM we will not show them to the user. We
do, however, want to export the data to the CRM still (as this
may make partial updates more straight forward in the future).

In order to do this a new `hidden?` state has been introduced;
if a step is `hidden?` it will not be shown to the user but
the pre-filled data _will_ be exported back to the CRM.