Consolidate hidden/skipped states
Instead of having a separate `hidden?` state - which is not
shown to the user but is exported - we combine it with the
`skipped?` state and change so that attributes are exported
from skipped steps using the existing CRM data.

This is going to be useful because skipped step values will be
`nil` for new candidates and will contain the data already in
the CRM for existing candidates. Ultimately this will make it
easier to update the API to support sending `nil` values to
the CRM (it avoids having to duplicate the logic for 'was this
step shown to the user' in the API).