Log address postcodes
We are seeing an issue where a postcode passes the Ruby validation but fails the API validation. This shouldn't happen as both use the same regex. Logging out the postcode will give us an example we can use to diagnose what's going on the next time this happens.

This shouldn't constitute personally identifiable information:

> The guidance is that Postcode will only count as PII if combined with other information which is ‘in the possession of, or is likely to come into the possession of, the data controller’ to identify a living individual.