Support accordions in the content layout
The accordion and content pages are so similar that it doesn't really
make sense to have two layouts, this change allows us to render
accordions beneath content on content pages. The next steps will be to
remove the accordion layout directives in the steps frontmatter and then
to delete the accordion layout.