Add temp logging to diagnose postcode validation error
We are seeing an error where an invalid postcode is being submitted. I suspect
this is an invalid postcode against a candidate already in the CRM, which is
then pre-filling the store in the app and being submitted _back_ to the API on
completion (which picks it up as a validation error).

Add temp logging to see if this is in fact the case.