Add parent/ancestor methods to Page model
To construct breadcrumbs we need to be able to traverse the content page
hierachy. This commit adds helper methods to get the parent of a `Page` as well
as all ancestors.

Due to the special nature of the home page the `ancestors` will return empty
for both the home page and other top-level pages. We will need to deal with the
home page use-case outside of this modelling when constructing the crumbs.