Render BreadcrumbComponent in layouts
This was a bit of a faff due to the style of the breadcrumbs changing depending
on if a hero section is rendered. We can't include the breadcrumbs in the hero
section as we sometimes don't render a hero at all (and other times we call to
render the hero component but it doesn't render as no frontmatter image is
supplied).

I'll follow this up with a clean up PR once merged as I'm sure it could be
tidier but will require further refactoring.