Render PageNotFoundError if markdown content missing
Currently we return a valid `Page` model for markdown content that does not
exist; Rails handles rendering the 404 when it finds the template of the `Page`
does not exist.

Instead, we now raise a `PageNotFoundError` and explicitly rescue it in the
`ApplicationController` to render 404 not found. We need to do this so that we
can more easily put in place methods to traverse the content hierachy when it
comes to building breadcrumbs.