Add methods to support traversing page hierachy
To build breadcrumbs of the content pages we need to be able to traverse the
page hieracy (from the request page up to the root).

Add `parent` method to get the parent of a page; a path may exist without
content so we skip parents that don't exist.

Add `ancestors` method to retrun all ancestors of the current page. These will
be the breadcrumbs.