Remove no-hero class
This was added so that we can apply a margin above the content when the hero
section is not present. It will no longer be needed as the
`BreadcrumbComponent` will add the relevant margin in.

The only place we still need to add a top-margin in is in the wizard completion
steps; these have a title that would otherwise overlap the logo.