Filter additional parameters
When we start logging the API client request/response models they will contain
camel case versions of attributes we should filter (firstName, lastName and
timedOneTimePassword).

As we are going to use the global `filter_parameters` to sanitize before
logging, this commit updates the list to include the camel case versions.