Optimise the placeholder replacer
The previous approach used to test the proof of concept requires
scanning through the text many times. Instead we can do it once and
replace as we go by passing a block to #gsub.

Unfortunately with Ruby, when using this approach the value passed to
the block argument seems to always contain all matching text from the
regexp, not just what's within the capture group.

I can't find a nicer way of referencing the captured text than using the
Perl backref style, hence the rubocop fence 😢