Make the right column (aside) content sticky
This makes the content, usually jump links or a CTA, follow the user as
they scroll down the page. We can amend this to _only_ make it work for
jump links if the CTA is unwanted.

It's only enabled on desktop and should work in all non-IE browsers,
they'll have to scroll up to the top as they currently do.