Add article_classes helper method
The article_classes method returns the classes that will be set on the
article element which surrounds most content on the site. By default
it'll just return `markdown`, but can append `fullwidth` when the
`full_width` key is set in the frontmatter.

Additionally, it can add arbitrary classes set via `article_classes`.
The intended use is to allow special classes to be used in certain
unusual situations, like on guidance pages or disclaimers.