Add support for title captions
Captions are a GOV.UK Design System pattern and they're useful for
adding a bit more context to headings without them getting too long.

This one will only be included when there's a `title_caption` in the
front matter and the CSS will place it just above the title while
maintaining the margin above.

https://design-system.service.gov.uk/styles/typography/#headings-with-captions