Allow mailing list sign up if already signed up for TTA
Previously we blocked someone signing up for the mailing list if they had
already signed up for a TTA (as having a TTA meant they were further along in
their journey/the mailing list wouldn't provide value).

As it turn s out, candidates (especially students), engage with an EEA in years
1 and 2. EEAs want to push them to the mailing list so they are kept engaged,
so we need to unblock that journey.