Standardise the way nav lists are built
Previously there was an inconsistency between the way the desktop and
mobile lists were built, the desktop variant used the #nav_link helper
and the mobile one didn't.

Also, instead of visiting one of the dynamically-added nav pages to
conduct the tests, visit the home page instead. This will ensure none of
the linked pages fall foul of #link_to_unless_current, which is used by
the #nav_link helper.