Update GroupPresenter to handle pagination
We want to be able to paginate the events sorted by type if they are part of
search results (as opposed to the upcoming events).

Add `paginated_events_by_type` method that paginates each set of events
according to the arguments (`pages_by_type` contains the active page for each
type and `per_page` is how many events per page we want to display).

Add `paginated_events_of_type` convenience method so that we can update
`EventCategoriesController` to use the presenter for pagination as well.