Fix for performance regression when running specs
The test suite was suddenly taking 5 minutes to run. This was caused by the view
paths list growing everytime "prepend fake views" was used ending up with 54
copies of that in the view paths.

The Frontmatter class was scanning each of these 54 duplicate paths in turn for
each of the tests relying on it.

This would only have affected test runs because production builds a preloaded
list on boot.