Ensure query string is maintained when redirecting
Rails' redirect method[0] will redirect entirely when passed a single
path argument. When passed a keyword argument for a specific part of the
URL - e.g., `path:`, only that part is substituted and the rest of
the URL is left as-is.

Here we use that functionality to our advantage and replace only the
path, this keeps the query params in place and should allow our tracking
to be maintained even when redirecting legacy pages to their new
counterparts.

[0] https://api.rubyonrails.org/v6.1.0/classes/ActionDispatch/Routing/Redirection.html#method-i-redirect