Update event search radius options
We have observed that nobody is really using the 50/100 mile radius search
options and the 30 mile radius option returns a decent number of results for
most people. Working on that assumption we want to make the event search more
localised and monitor it going forward, in hopes that users find the tighter
search radius more useful.