Ensure the 404 is rendered in HTML
We've seen an influx of bots poking around for pages that don't exist.
These cause a ActionView::MissingTemplate error because we're trying to
serve a 404 in the format they requested the docuemnt in, so
`/some-file.md` is expecting us to have a `errors/not_found.md` file,
which we don't.

Instead, force the 404 to HTML.