Add metrics for page speed scores
We want to record the page speed score for performance, accessibility and seo
in Prometheus.

Adds the metrics as gauages (as they will go up/down) with labels for strategy
(mobile/desktop) and path (page under test).

The path label will have a fairly high cardinality, but it is finite (the
number of pages on the site) and will give us page-level insights.