Add PageSpeedScore class
Retrieve page speed scores via the google-api-client for each page listed in
the `sitemap.xml`, querying performance, accessibility and seo categories for
both mobile and desktop strategies.

The fields have been restricted to make what is a slow request as fast as
possible and the pages are queried asynchronously using up to 10 threads at a
time.