Cache scores and publish metrics on application start
Initially we intended to use the Prometheus Push Gateway to send the metrics
directly from the Rake task that retrieves the page speed scores. It turns out
that it would be difficult to secure the push gateway, so we need to use a
different method.

In this commit the scores are written to Redis from the Rake task and when the
application boots in the rolling environment the scores are published for
Prometheus to scrape.