Fingerprint content assets
We are not currently able to have an efficient caching strategy on the assets
in the content repository due to them not having fingerprints.

By adding fingerprints to these assets we can update the Rails static cache to
be much longer than the current 15 minutes (we could change it so that the
assets get cached forever).

Add a `Fingerprinter` class that fingerprints the assets in the given
`asset_dir` and updates references to them in the given `content_dir` (inspired
by how Middleman's `:asset_hash` extension).

Add a rake task to run the fingerprinting against the content repository
assets/content, which could be called as part of the Docker build process:

```
bundle exec rake fingerprinter:run
```

If this gets merged down the change to the content repo Dockerfile can be made
and it can all be tested in the generated review app.

The matching is fairly liberal given all the weird and wonderful ways you can
link to an image between markdown and ERB. A side-effect of this is that if the
content text contains the path to a valid image it will also get the
fingerprint added in - that may or may not be desirable, but its hard to avoid
and I can't see a time when we would display the path to an asset directly to
the user, so I haven't tried to avoid it for now.