Add NextGenImages class
Introduces a class that replaces `<img>` tags with `<picture>` tags and
applicable `<source>` tags for any next-gen images that exist with the same
file name as the original image.

This will allow us to generate next-gen images as part of the Docker build
process and then Rails will introduce them into the responses dynamically. As
most of our pages are static and cached it shouldn't have a big performance
overhead.