Support lazy loading images that are below the fold
Intercepts HTML before rendering and removes the `src` attribute, retaining the
path in `data-src`.

Appends a `noscript` tag containing the original `img` to support when
Javascript is disabled.

The original images are hidden by default, to avoid styling issues when JS
disabled (we don't want to show both images as the lazy image will not have a
src set).

Lazysizes picks up any img tags with a `lazyload` class and populates the `src`
when the user is scrolling towards the image. It also supports lazy-loading the
picture source tags with `data-srcset`.

A side-effect of this is that the webp/jp2 image sources will no longer be
available if the browser has JS disabled. Whilst we could fix this I don't
think its worth the effort given the % of users in this scenario.

Enabled only in preprod environment for now.