Conditionally enable VWO on pages
By default we want to disable VWO on all pages, as we are no longer using it
for redirections/splitting traffic.

We will be using VWO in the future to run a/b tests, however, so its useful to
keep it around but have a mechanism to enable it on specific paths.

Updates `AnalyticsController` to work when VWO is disabled.