Run the page speed task on boot
We now have a page speed environment running alongside preprod/dev. Instead of
figuring out how to run a rake task in Terraform and reboot the instance so
that it displays the metrics to Prometheus I've updated the app to run the page
speed tool when the worker boots.

We have a new set of credentials to ensure the app is booted with
`RAILS_ENV=pagespeed`.

Removes the rake task.