Hide original images when JS disabled
When JS is disabled the original `img` tag would be `visibility: hidden` but as
it has an explicit width/height it still takes up space (if we use `display:
none` it would result in content shift when the image is set and the selector
no longer applies).

Instead, this adds a new pack with CSS for when JS is disabled that will
`display: none` on the original image (as JS is disabled the content shift will
never happen as the lazy image loading will be working and a default noscript
img is shown).