Improve IE11 suppprt with babel polyfills
Remove the manual polyfills.

Add browserlist.rc so that `present-env` can generate appropraite polyfills
when transpiling.

Import `core-js` and `regenerator-runtime/runtime` as `@babel/polyfill` is now
deprecated.

The benefit is we don't need to worry about polyfilling IE manually going
forward; but the drawback is it adds about 70kb to the application pack size.