Smooth transformed elements on mobile Safari
On iOS mobile Safari the rotated blocks end up with pixelated edges due to the
use of `backdrop-filter: blur()`. I can't find a way of retaining the blur and
anti-aliasing the edges, so removing it only for mobile Safari (the blur is a
nice to have and was more important when we had rotated blocks that mostly
overlapped images, which I don't think we have any more).