Add some basic font scaling utilities
Here we're adding two smaller fonts to our map and for every size we go
down (ie, from desktop to tablet) we're shifting everything in the map
one to the left.

Currently it's done manually rather than in an indexed loop - we might
need a little more control here or there but that could be a future
improvement.