Update API cache store
Currently, when a teaching event is posted to the API, it will not be immediately available on the hosted environments because of HTTP caching.

The cache will be cleared when events are posted in the API client Gem. However, we only want to clear the API cache, so here we update the API client cache store to use a new Redis namespace.