Run specs in random order
At the moment if we hit a flakey spec in the CI pipeline we have no way to
recreate the same test order locally in order to debug it.

By running in a random order RSpec will print out the `seed` value used for
the run, which we can then use to perform an identical run locally.