Re-enable turbolinks preview cache on events layout
This was originally disabled due to it not behaving with the pagination
component, but its causing an issue where the page is skipped completely when
you press the back button (so hitting back on an event page goes to the home
page and not the find events page). Re-enabling it I can't replicate the issue
with the pagination anyway.