Support responsive images
We want to be able to display different images depending on the device.

Add `ResponsiveImages` lib that will generate the appropriate `source` tags for
any responsive images that it finds present in the filesystem. The images must
take a specific format to be used:

```
/path/to/image--<breakpoint>.ext
```

If a responsive image is found, it will be displayed for the device specified
in the `breakpoint` and all smaller viewports (unless a specific image exists
for a smaller breakpoint):

```
mobile: "500px"
tablet: "768px"
desktop: "1000px"
wide: "1500px"
```