Add rel='noopener noreferrer' to outbound links
This is probably not a risk for users of newer browsers but adds a layer
of security for users of out-of-date ones. It can prevent the pages we
link to from gaining 'partial access' to our page; this could allow an
attacker who has gained control of a page we link to to cause mischeif
by modifying the `window.opener.location`, etc.

More info here:
* https://dev.to/ben/the-targetblank-vulnerability-by-example
* https://medium.com/@jitbit/target-blank-the-most-underestimated-vulnerability-ever-96e328301f4c#.disn238f9