Remove JS `link_controller`
The link controller creates a hidden node for outbound links. The most recent change to the link controller targets all a tags.

This is not working well with the Trix editor, where new links are being created for every link in that field. These appear as: (Link opens in new window) when an event is duplicated or edited. Removing the link controller from the internal layout to fix.