Move git-font-family declaration above git import
There were a few places, like in the 'All content is available under'
footer paragraph, where GOV.UK font settings were being applied. Because
GDS Transport isn't availble it was defaulting to a sans font, which
(for some reason, on my machine at least) was different to the sans
font we use elsewhere.

Moving the $git-font-family declaration above the import of the 'git'
stylesheet prevents this from happening.