Split internal styles out to their own pack
This should provide an extra level of safety from the internal files
affecting the rest of the site.

One very-slight side effect was seen when some govuk classes were
causing the non-standard font family to be used because the internal
stylesheet imports all of govuk-frontend whereas the git stylesheet
expects only some parts of it.

It'll also make the page a bit lighter for most visitors.