Update search with ranking/levenshtein comparisons
In order to improve the search we want to rank the results (more keyword
matches resulting in a higher-ranked page), make use the Levenshtein comparison
to reduce spelling errors impacting the search results and remove stopwords to
reduce noise.

The search algorithm now consists of:

- Tokenize the search term, keywords and page title (removing stopwords, downcase)
- For each page
  - Count keyword matches and multiply by 2 to add weight
  - Count page title word matches
  - Sum to give page rank
- Remove unranked pages (no matches)
- Sort pages by rank, descending

A 'match' is defined as:

- The word starts with one of the search tokens
- The word is > 5 characters and Levenshtein distance is <= 1 to a search token

The original search tests have been left as unchanged as possible to show that
the new search works with all the original test cases. Tests for the new search
logic will be in a follow up commit.