Make sure that users cannot enter a year with more than 4 digits
When entering a date manually, rather than using the selector, it is possible to enter a 6 digit year. This slows the user down when filling out the form quickly. 

Add a max year attribute, so that the input now moves the user on to the time input after the date is completed.