Rename accept cookies button id to bypass ad-blockers
It appears that AdBlockers in general allow our cookie modal to appear, however
they hide the "Accept all cookies" button (seemingly targetting the `id`
because it includes `cookie`).

A workaround is to rename the `id` to something else - I've left a note in the
code to prevent this from being changed back in the future.