Introduce the Pages::Navigation model
It's now responsible for building up a collection of nodes
(Pages::Navigation::Node) that each represent a page on the site in the
navigation tree, and is used to build the existing navigation (both
desktop, mobile and footer).

Currently we're only interested in root pages, but this structure will
make it easier to find child pages for a particular node.

Unlike the previous iteration, we're now leaning towards using the path
for hierarchy and determining whether or not a page is root based on it.
This is going to be easier to understand than setting `parent` metadata
in the front matter.