Introduce delay before executing search
Currently the search is executed after every key press, which both results in a
lot of search API calls and also means our analytics include search terms for
each character.

Instead, this introduces a 500ms delay after the user finishes typing before
executing the search.

It would have been nicer to use Jest's fake timers for the tests, however the
`accessible-autocomplete` lib uses timers to poll the input field internally
and this makes using the fake timers not possible (it just hangs, even when you
ask it to only run on the pending timers).