Update to show 'Searching...' while searching
The no results text is displayed while searching by default; as the search has
a 500ms delay we want to show 'Searching...' until the request has completed.
We can achieve this by changing the no results text dynamically depending on
the search state.

I wasn't able to test the 'No results found' text still displays correctly in
Jest due to using the XMLHttpRequest, but I've tested it manually and it works
fine. We should probably swap this out for `fetch` at some point.