Add check for static asset availability
We are seeing sporadic 404s being returned for assets that contain a digest
from an old deployment. We suspect CloudFront is not caching these assets
according to the 1 year cache-control TTL we have set, but we want to determine
if users are running into this error during deployments when the asset digests
are changing, or if its just bots poking about.

This introduces an `AssetChecker` that loads the home page, pulls out the
CSS/JS pack URLS and then attempts to download them; if the requests fail it
will be returned as a 404 to the client with a list of the failed URLs, if it
passes then we return an empty 204.

We are using the page speed environment for ease; this doesn't really warrant
its own environment and it can be torn out once we get to the root cause of the
404ing.