Change to POST request
Its showing as an unauthorized request in the GitHub Actions, I'm not sure why
as we send the username/password in the same way as the page speed task. The
only difference I can see is one is this is a GET request; changing to a POST
to see if that makes a difference.