Prevent double clicking form buttons
As we are not using the `govuk_submit` helper (due to our own submit button
styling) the default prevent double click behaviour was not being applied.

Update to include the `data-prevent-double-click` and `data-disable-with`
for correct double click prevention behaviour.

Update disabled button styling to fade it out slightly so it looks
disabled.