Set `is_online` to true for online type events on creation
All Online type events must have the is_online property set to true. Currently, there is a potential for Online type events to be posted to the API with is_online set to nil. An event object is always initialised with a type_id, so set is_online there.

The behaviour for Provider events will remain the same, where the user chooses via the form.