Fix the external link checker
The external link checker has been failing on the
nightly runs for a couple of reasons.

Some websites don't respond to HEAD requests; we
have a mechanism in place already to perform GET
requests for specific websites, but that requires
us to manually update a list; a new link that
fails HEAD requests was added and the list wasn't
updated. Instead, I've changed it to try a HEAD
request and fallback to a GET request dynamically
if it fails (by either returning a 404 or there
being a connection error).

A number of other external links also appear to
fail Faraday's SSL certificate checks. I don't
think these are bullet-proof as I checked a few
manually and they were fine, so I've had to add
these domains to the ignore list for now.