Allow users to paste a date and time into `start at` and `end at` inputs
The users of the `internal/events` forms receive events via email from the Advertise a Teaching Event form on the main website. They would like the ability to paste a date and time into the `start at` and `end at` fields in the format of the email (e.g., `01/07/2021 17:00`), as it is tedious to type the full date out or to use the selector. This updates the `internal/events` form to use `flatpickr` for its datetime fields, which lets us paste into the input and specify custom formatting.

The API accepts a full ISO format datetime such as `2021-07-01T09:00:00.000Z`, or a localised format such as `2021-07-02T09:00:00`. However, the CRM seems to store the localised version without any additional time zone information (it also stores a separate time zone code), which in turn will update the API cache after a sync. This means (without changing how the API works) it is easier to work with a localised datetime in the APP. Otherwise, the format of the returned date will unexpectedly 
change when a sync runs, and both formats will need to be handled.

We were previously using `<input type="datetime-local">` which also stores a localised datetime with no timezone information and was working nicely.