(Hopefully) fix flakey rate limit specs
These appear to fail completely at random - an rspec bisect results in a test
sequence that subsequently passes.

My theory is that `Rack::Attack` works on time boundaries rather than a rolling
window; so I expect if the test runs across a minute boundary and the period is
also 1 minute then it will fail because it expects a 429 and gets a 200. To
avoid this I am freezing time at the start of the test. I've also added more
padding to the travel time to check the rate limit doesn't apply after the
period has elapsed (10 seconds up from 1 second) incase there's something iffy
going on there as well.