Set to London timezone on callback booking
The book a callback form displays time slots from the API,
which are in UTC time. We need to set the time zone explicitly
to `London` for this wizard flow in order to get the times
outputting with the correct offset (they started at 8am
previously when the time should show as 9am).