Add a similar posts method to Pages::Blog
Similarity is judged based on the number of overlapping tags. If there
is a tie the posts are just returned in their original order as we're
not forming an ordered list or ranking.

The original post is excluded from the results and a limit can be set,
currently it defaults to 3.