Add production credentials for internal events portal
The internal events portal can be used to create and publish teaching events. These are created in a pending state, and when published will appear on the main website.

Currently, the events team and CRM team have been using the portal to create pending teaching events in Dev before pushing them to Production through the CRM. 

They are now looking to start using the events portal in Production. This means that the entire process (creating and publishing events) can be done through the website.