Hard-code the mailing list URL
We're currently relying on Rails' routing for generating the mailing
list paths in these two places but using the hard-coded path everywhere
else in the app.

Due to the way the sign-up wizard is written the individual pages don't
have their own paths; instead they rely on a redirect in
WizardSteps#index to send users to the right place.

The fact that a _correct_ navigation to `/mailinglist/signup` results in
a 302 and the user is routed to `/mailinglist/signup/name` (the first
step in that particular wizard) means there's no specified route for
each individual page in the wizard, and seemingly no _simple_ way of
generating a path with a helper (eg. `mailing_list_steps_path("name")`)

We believe this redirection has had an impact on paid search results.

We have two options here:

* Supply the wizard class to the `step_path` method so we can use Rails'
  built-in functionality:

  ```
    redirect_to step_path(wizard_class.first_key, query_params)
  ```

* or, hard-code it

I'm going for the latter here as it's lighter touch, we don't need to
introduce any new routing rules or helpers to ease the process and
we can test it out. It also makes links to the mailing list easier to
find via ag/grep

The disadvantage is that we can't change the routing on a whim and have
the new rules reflected throughout automatically.