Load all future events in category view
When we are viewing an event category we only want to allow the user to filter
over the next 6 months, but give them the option to page through the entire
list of events.

As the to/from date logic is reasonably complex already, the simplest solution
is to extend the future date from 6 months to 24 (which should cover all future
events) and filter down the `available_months` in the `SearchComponent`.