Improve the radio button size on Chrome/Safari
The aspect-ratio didn't appear to be having an affect on the input in
Chrome and Safari, it was stretching.

Now the height and width are forced at 2em. Also, because the inherited
form style uses outline instead of border for the focus state, Chrome
was making a square box around focussed radio buttons; we're now
reverting to a border.