Only show teaching logo for first category
We show the teaching logo for any category that is not School and
university events. This has been fine whilst we haven't had any Train
to Teach events, but now we are getting some back on the website it
means both categories end up with the teaching logo.

Instead, we want to show the teaching logo for only the first
category that is _not_ School and University events.