Track number of failed matchback attempts
We want to track the number of times a user fails to matchback as part of the
`IssueVerificationCode` concern, so that we can have custom copy on the book a
callback feature for candidates that are failing matchback.

We are also tracking the last failure code from the CRM so that we can tailor
the copy depending on if the user wasn't found vs the CRM is unavailable.