Update copy for when there are no TTT events
Make it clearer that there are no Train to Teach events for your chosen month,
but there may be for other months. The previous message leads the user to
believe that TTT events won't be added until September, which isn't the case -
they may be availeble now for the September time period.