Allow searching by all months
We default to the curent month in the search; this has been fine but now we
have TTT events that start in a few months time and we find users click the
search with the default month values most often. So that TTT events are in the
results we are changing the month selector to default to 'all months'.