Change to a sizing scheme not based on px
Previously the 'basis' for our font sizing was whole numbers from 14-64,
where the number represented the size in pixels.

This left us with little room for manouvre when it comes to making fine
adjustments to sizes.

Now we simply have a list of sizes and and we pick a different range
from the list for each screen size.

There is almost-certainly a nicer way of doing this in Sass, using
nested loops, but this implementation is straightforward.