Native support for Question Time events
Currently the GiT API remaps Question Time events with a Train to Teach
event type. This is because the events team want the QT and TTT events both
displayed as if they are TTT events on the GiT website.

This, however, is a bit of confusing and potentially hazardous code in the
GiT API and we would prefer to remove this and support QT events natively
in the GiT website.

This commit achieves that by:

- Updating the `Events::GroupPresenter` to group TTT and QT events for
  displaying in the event search results.

- As a byproduct of this ^ we need the presenter to now know about how many
  events there should be per-type (as we're combining so could end up with
more than expected).

- Updating the `Events::Search` model to remove QT events as a searchable
  option and query both TTT and QT events if searching for TTT events.

- Mis-representing QT events as TTT event types in the event box by
  providing translations.

The API was previously updated to support querying by multiple `type_ids`
so that when searching we can perform a single search for both QT and TTT
events.