Compile assets once per test run
According to Stackprof the test run is constantly computing digests for the
assets. The PR to move all our assets into webpacker caused the runtime of the
tests to double. Instead of checking constantly for asset recompilation we can
disable this in the test environment and run a single `Webpacker.compile` for
the suite. This brings the test run time from 5m back down to around 2m.